backend(config): adds config module

+++ apps/server/src/library/library.service.ts
@@ -1,15 +1,17 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateLibraryDto } from './dto/create-library.dto';
 import { UpdateLibraryDto } from './dto/update-library.dto';
 import LibraryRepository from './repository/library.repository';
 import Library from './schema/library.schema';
 import FileService from 'src/utils/files/file.service';
+import { ConfigService } from '@nestjs/config';
 
 @Injectable()
 export default class LibraryService {
   constructor(
     private readonly repository: LibraryRepository,
     private readonly fileService: FileService,
+    private readonly configService: ConfigService,
   ) {}
   create(createLibraryDto: CreateLibraryDto): Promise<Library> {
     return this.repository.create(createLibraryDto);
@@ -19,31 +21,39 @@ export default class LibraryService {
     return `This action returns all library`;
   }
 
-  findOne(id: string): Promise<Library> {
-    return this.repository.findById(id);
+  find(id: string): Promise<Library> {
+    const doc = this.repository.findById(id);
+    if (doc) {
+      return doc;
+    }
+    throw new NotFoundException();
   }
 
   update(id: string, updateLibraryDto: UpdateLibraryDto): Promise<Library> {
     return this.repository.findByIdAndUpdate(id, updateLibraryDto);
   }
 
-  remove(id: string): Promise<null> {
-    return this.repository.findByIdAndDelete(id);
+  async remove(id: string): Promise<void> {
+    // 1. Fetch doc from db
+    const fileDoc = await this.find(id);
+
+    // 2. Delete file for s3
+    const key = fileDoc.path;
+    const res = await this.fileService.delete(key);
+
+    // 3. If successfully deleted from s3 then delete from db
+    if (res) {
+      await this.repository.findByIdAndDelete(id);
+    }
   }
 
   async upload(file: Express.Multer.File, path: string) {
-    const s3FilePath   = await this.fileService.upload(file, path);
+    const s3FilePath = await this.fileService.upload(file, path);
     const filePayload: CreateLibraryDto = {
       name: file.originalname,
       type: file.mimetype,
-      path: s3FilePath.path
+      path: s3FilePath.path,
     };
     return await this.create(filePayload);
-  }
-
-  async delete({key, id}: {key: string, id: string}) {
-    const destinationBucket = 'dokku-assets';
-    await this.fileService.delete(key, destinationBucket);
-    return await this.remove(id);
   }
 }
