+++ apps/server/src/notes/notes.service.ts
@@ -3,8 +3,9 @@ import { CreateNoteDto } from './dto/create-note.dto';
 import { UpdateNoteDto } from './dto/update-note.dto';
 import NotesRepository from './repository/notes.repository';
 import Note from './schema/notes.schema';
-import { SchemaType, SchemaTypes, Types } from 'mongoose';
+import { Types } from 'mongoose';
 import FileService from 'src/utils/files/file.service';
+import SaveNoteDto from './dto/save-note.dto';
 
 @Injectable()
 export default class NoteService {
@@ -13,9 +14,8 @@ export default class NoteService {
     private readonly fileService: FileService,
   ) {}
   async create(createNoteDto: CreateNoteDto): Promise<Note> {
-    const res = await this.upload(createNoteDto, 'temp');
-    
-    return this.repository.create({...createNoteDto, path: res.path});
+    const res = await this.upload({ ...createNoteDto, path: 'temp' });
+    return this.repository.create({ ...createNoteDto, path: res.path });
   }
 
   findAll() {
@@ -23,41 +23,52 @@ export default class NoteService {
   }
 
   async find(id: string): Promise<Note> {
+    // 1. fetch doc from db
     const doc = await this.repository.findById(id);
     if (doc) {
+      // 2. fetch content from s3
+      const key = doc.path;
+      const res = await this.fileService.getFile(key);
       return doc;
     }
     throw new NotFoundException();
   }
 
-  async update(id: string, updateNoteDto: UpdateNoteDto): Promise<Note | void> {
-    //step 1: if content is present
-    //step 2: if content is present then upload to s3
-    //step 2.1: fetch path from db
-    //step 2.2:
-    //step 3: else update db
-    if (updateNoteDto.content) {
-      const doc = await this.find(id);
-      const noteData = {
-        name: doc.name,
-        content: updateNoteDto.content,
-      };
-      await this.upload(noteData, doc.path);
-      return;
-    } 
+  async update(id: string, updateNoteDto: UpdateNoteDto): Promise<Note> {
     return this.repository.findByIdAndUpdate(id, updateNoteDto);
   }
 
-  remove(id: string): Promise<null> {
-    return this.repository.findByIdAndDelete(id);
+  async saveNote(saveNoteDto: SaveNoteDto, id: string): Promise<string> {
+    const doc = await this.find(id);
+    const noteData = {
+      name: doc.name,
+      content: saveNoteDto.content,
+      path: doc.path.split('/').slice(0, -1).join('/').replace(/\/$/, ''), //path -> userId/notes/file_name.json
+    };
+    await this.upload(noteData);
+    return;
   }
 
-  async upload(noteData: CreateNoteDto | UpdateNoteDto, path: string) {
+  async remove(id: string): Promise<void> {
+    // 1. Fetch doc from db
+    const fileDoc = await this.find(id);
+
+    // 2. Delete file for s3
+    const key = fileDoc.path;
+    const res = await this.fileService.delete(key);
+
+    // 3. If successfully deleted from s3 then delete from db
+    if (res) {
+      await this.repository.findByIdAndDelete(id);
+    }
+  }
+
+  async upload(noteData: { name?: string; content: string; path: string }) {
     const notePayload = {
       name: noteData?.name ?? new Types.ObjectId().toString(),
       type: 'json',
       content: Buffer.from(noteData.content, 'utf-8'),
     };
-    return await this.fileService.uploadJSON(notePayload, path);
+    return await this.fileService.uploadJSON(notePayload, noteData.path);
   }
 }
